Number of small fixes and improvements
UI
- added "About" dialog to "Tools" button (in package dialog)

Fixes for VS2013 compilation
- nvtt/nvcore/DefsVcWin32.h - va_copy define there caused "macro redefinition" warning
- common.project - added _USING_V110_SDK71_ define to avoid similar warnings in other files
- UIProgressBar - avoid use of float to int converter which is missing in msvcrt

Tools:
- genmake now adds DEFINES for resource compiler (rc.exe) command line
- CompatTable - fixed layout of "filter" input

+++ UI/BaseDialog.cpp
@@ -534,7 +534,7 @@ void UIProgressBar::SetValue(float value)
 {
 	if (value == Value) return;
 	Value = value;
-	if (Wnd) SendMessage(Wnd, PBM_SETPOS, Value * 16384, 0);
+	if (Wnd) SendMessage(Wnd, PBM_SETPOS, (int)(Value * 16384), 0);
 }
 
 void UIProgressBar::Create(UIBaseDialog* dialog)
@@ -543,7 +543,7 @@ void UIProgressBar::Create(UIBaseDialog* dialog)
 	Parent->AllocateUISpace(X, Y, Width, Height);
 	Wnd = Window(PROGRESS_CLASS, "", 0, 0, dialog);
 	SendMessage(Wnd, PBM_SETRANGE, 0, MAKELPARAM(0, 16384));
-	if (Wnd) SendMessage(Wnd, PBM_SETPOS, Value * 16384, 0);
+	if (Wnd) SendMessage(Wnd, PBM_SETPOS, (int)(Value * 16384), 0);
 }
